fix: editable state post drag
